Refatorando (atualizando código)

+++ src/main/java/com/educandoweb/course/resources/UserResource.java
@@ -2,8 +2,11 @@ package com.educandoweb.course.resources;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,22 +19,27 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import com.educandoweb.course.entities.User;
+import com.educandoweb.course.repositories.UserRepository;
 import com.educandoweb.course.services.UserService;
+import com.educandoweb.course.services.exceptions.DatabaseException;
+import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @RestController
-@RequestMapping(value = "/users")
+@RequestMapping("/users")
 public class UserResource {
 
 	@Autowired
 	private UserService service;
 	
+	@Autowired
+	private UserRepository userRepository;
+	
 	@GetMapping
-	public ResponseEntity<List<User>> findAll(){
-		List<User> list = service.findAll();
-		return ResponseEntity.ok().body(list);
+	public List<User> findAll(){
+		return userRepository.findAll();
 	}
 	
-	@GetMapping(value = "/{id}")
+	@GetMapping("/{id}")
 	public ResponseEntity<User> findById(@PathVariable Long id) {
 		User obj = service.findById(id);
 		return ResponseEntity.ok().body(obj);
@@ -44,15 +52,34 @@ public class UserResource {
 		return ResponseEntity.created(uri).body(obj);
 	}
 	
-	@DeleteMapping(value = "/{id}")
+	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> delete(@PathVariable Long id) {
-		service.delete(id);
-		return ResponseEntity.noContent().build();
+		try {
+			if (!userRepository.existsById(id)) {
+				return ResponseEntity.notFound().build();
+			}
+			
+			service.delete(id);
+			
+			return ResponseEntity.noContent().build();
+		} catch (DatabaseException e) {
+			return ResponseEntity.status(HttpStatus.CONFLICT).build();
+		} catch (ResourceNotFoundException e) {
+			return ResponseEntity.notFound().build();
+		}
 	}
 	
-	@PutMapping(value = "/{id}")
-	public ResponseEntity<User> update(@PathVariable Long id, @RequestBody User obj) {
-		obj = service.update(id, obj);
-		return ResponseEntity.ok().body(obj);
+	@PutMapping("/{id}")
+	public ResponseEntity<?> update(@PathVariable Long id, @RequestBody User user) {
+		Optional<User> userAtual = userRepository.findById(id);
+
+		if (userAtual.isPresent()) {
+			BeanUtils.copyProperties(user, userAtual.get(), "id");
+
+			User userSalvo = service.insert(userAtual.get());
+			return ResponseEntity.ok(userSalvo);
+		}
+
+		return ResponseEntity.notFound().build();
 	}
 }
